+++ analysis/src/main/java/cz/vutbr/fit/group/verifit/oslc/analysis/automationPlans/AutomationPlanConfManager.java
@@ -10,8 +10,19 @@
 
 package cz.vutbr.fit.group.verifit.oslc.analysis.automationPlans;
 
+import java.io.File;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Set;
+
+import org.eclipse.lyo.store.StoreAccessException;
+
+import cz.vutbr.fit.group.verifit.oslc.analysis.outputParser.IParser;
+import cz.vutbr.fit.group.verifit.oslc.analysis.outputParser.ParserManager;
+import cz.vutbr.fit.group.verifit.oslc.analysis.properties.VeriFitAnalysisProperties;
 
 /**
  * A singleton class responsible for holding AutomationPlan configuration properties that are not included in the
@@ -21,10 +32,14 @@ public class AutomationPlanConfManager {
 
     private static AutomationPlanConfManager INSTANCE;
 
+    private AutomationPlanLoading autoPlanLoader;
+    private ParserManager parserManager;
     private Map<String, AutomationPlanConf> automationPlanConfigurations;
-
+    
     private AutomationPlanConfManager() {
         this.automationPlanConfigurations = new HashMap<String, AutomationPlanConf>();
+        this.autoPlanLoader = new AutomationPlanLoading(new File(VeriFitAnalysisProperties.AUTOPLANS_DEF_PATH));
+        this.parserManager = new ParserManager(Paths.get(VeriFitAnalysisProperties.PLUGIN_PARSER_CONF_PATH));
     }
 
     public synchronized static AutomationPlanConfManager getInstance() {
@@ -35,26 +50,44 @@ public class AutomationPlanConfManager {
         return INSTANCE;
     }
 
+    public void initializeAutomationPlans() throws StoreAccessException, Exception
+    {
+    	Collection<AutomationPlanConf> autoPlanConfs = this.autoPlanLoader.loadAutomationPlans();
+    	this.parserManager.loadParsers(autoPlanConfs);
+    	for (AutomationPlanConf conf : autoPlanConfs) {
+    		this.automationPlanConfigurations.put(conf.getIdentifier(), conf);
+    	}
+    	this.autoPlanLoader.persistAutomationPlans(this.automationPlanConfigurations);
+    }
+    
     public AutomationPlanConf getAutoPlanConf(String identifier)
     {
         return this.automationPlanConfigurations.get(identifier);
     }
 
+    public Set<String> getAllAutoPlanIds()
+    {
+        return this.automationPlanConfigurations.keySet();
+    }
+    
     public void addAutoPlanConf(String identifier, AutomationPlanConf conf)
     {
         this.automationPlanConfigurations.put(identifier, conf);
     }
 
-
     public static class AutomationPlanConf {
+    	private String identifier;
         private String launchCommand;
         private String toolSpecificArgs;
         private Boolean oneInstanceOnly;
+        private Map<String, IParser> parsers; 
 
-        public AutomationPlanConf(String launchCommand, String toolSpecificArgs, Boolean oneInstanceOnly) {
+        public AutomationPlanConf(String identifier, String launchCommand, String toolSpecificArgs, Boolean oneInstanceOnly) {
+        	this.identifier = identifier;
             this.launchCommand = launchCommand;
             this.toolSpecificArgs = toolSpecificArgs;
             this.oneInstanceOnly = oneInstanceOnly;
+            this.parsers = new HashMap<String, IParser>();
         }
 
         public String getLaunchCommand() {
@@ -68,5 +101,26 @@ public class AutomationPlanConfManager {
         public Boolean getOneInstanceOnly() {
             return oneInstanceOnly;
         }
+        
+        public String getIdentifier() {
+            return identifier;
+        }
+
+		public Map<String, IParser> getParsers() {
+			return parsers;
+		}
+
+		public IParser getParser(String name) {
+			return parsers.get(name);
+		}
+		
+		public Boolean containsParser(String name) {
+			return parsers.containsKey(name);
+		}
+
+		public void putParser(String parserName, IParser parser) {
+			this.parsers.put(parserName, parser);
+		}
+        
     }
 }