+++ build.ps1
@@ -46,6 +46,9 @@ function main ()
     echo "    Building and Installing shared resources"
     echo "############################################################"
     echo ""
+
+    mvn install:install-file -Dfile='.\lib\cz.vutbr.fit.group.verifit.jsem_0.2.0.202103021435.jar' -DgroupId='cz.vutbr.fit.group.verifit.jsem' -DartifactId='jsem' -Dversion='0.2.0.qualifier' -Dpackaging='jar'
+
     cd shared
     mvn clean install